function/requirements.txt
fastapi
os
azure-mgmt-compute
azure-identity
requests
azure-storage-queue
python-dotenv
azure-functions